Implement my question modification function

+++ src/main/java/cn/edu/tit/community/dto/QuestionDTO.java
@@ -14,6 +14,6 @@ public class QuestionDTO {
     private Integer likeCount = 0;
     private Long gmtCreate;
     private Long gmtModify;
-    private Long creator;
+    private int creator;
     User user;
 }
